rename /evidence/SomaticMutation to /evidence/OpenPedCanSomaticMutations
